final 2

+++ services/contactsServices.js
@@ -1,32 +1,36 @@
 import { v4 as uuidv4 } from 'uuid';
 import { Contacts } from '../models/contactsModel.js';
 
-async function listContacts() {
-    return await Contacts.find().catch(error => {
+async function listContacts(contactOwner) {
+    return await Contacts.find({ owner: contactOwner}).catch(error => {
         console.error(error);
         return [];
     });
 }
 
-async function getContactById(contactId) {
+async function getContactById(contactId, contactOwner) {
     try {
-      return await Contacts.findOne({ _id: contactId }) || null;
+      const contact = await Contacts.findOne({ _id: contactId }) || null;
+      if (contact.owner !== contactOwner) return null
+      return contact
     } catch (error) {
       return null;
     }
 }
 
-async function removeContact(contactId) {
+async function removeContact(contactId, contactOwner) {
     try {
-        return await Contacts.deleteOne({ _id: contactId });
+        const contact = Contacts.findOne({ _id: contactId });
+        if (contact.owner === contactOwner) return await Contacts.deleteOne({ _id: contactId });
+        return null
     } catch (error) {
       return null;
     }
 }
   
-async function addContact(name, email, phone, favorite) {
+async function addContact(name, email, phone, favorite, owner) {
     try {
-        const newContact = await Contacts.create({ id: uuidv4(), name, email, phone, favorite });
+        const newContact = await Contacts.create({ id: uuidv4(), name, email, phone, favorite, owner });
         return newContact;
     } catch (error) {
       return null;
